pundit updates

+++ app/controllers/admin/system/affiliates_controller.rb
@@ -18,14 +18,12 @@ class Admin::System::AffiliatesController < Admin::BaseController
   end
 
   def new
-    @affiliate = Affiliate.new(name: 'New affiliate', active: true, country: Cache.setting(:system, "Default Country"))
-    authorize @affiliate
+    @affiliate = authorize Affiliate.new(name: 'New affiliate', active: true, country: Cache.setting(:system, "Default Country"))
     render 'edit'
   end
 
   def create
-    @affiliate = Affiliate.new(affiliate_params)
-    authorize @affiliate
+    @affiliate = authorize Affiliate.new(affiliate_params)
     
     if @affiliate.save
       unless params[:c].blank?
@@ -39,18 +37,15 @@ class Admin::System::AffiliatesController < Admin::BaseController
   end
 
   def show
-    @affiliate = Affiliate.find(params[:id])
-    authorize @affiliate
+    @affiliate = authorize Affiliate.find(params[:id])
   end
 
   def edit
-    @affiliate = Affiliate.find(params[:id])
-    authorize @affiliate
+    @affiliate = authorize Affiliate.find(params[:id])
   end
 
   def update
-    @affiliate = Affiliate.find(params[:id])
-    authorize @affiliate
+    @affiliate = authorize Affiliate.find(params[:id])
     
     if @affiliate.update(affiliate_params)
       redirect_to action: 'show', id: @affiliate.id, notice: 'Affiliate was successfully updated.'
@@ -60,8 +55,7 @@ class Admin::System::AffiliatesController < Admin::BaseController
   end
 
   def destroy
-    @affiliate = Affiliate.find(params[:id])
-    authorize @affiliate
+    @affiliate = authorize Affiliate.find(params[:id])
     @affiliate.destroy
     
     redirect_to action: 'index', notice: 'Affiliate has been deleted.'
@@ -69,14 +63,18 @@ class Admin::System::AffiliatesController < Admin::BaseController
   
   def extra_properties
     @affiliate = Affiliate.find(params[:id])
+    authorize @affiliate, :show?
     5.times { @affiliate.extra_properties.build }
   end
   
   def categories
     @affiliate = Affiliate.find(params[:id])
+    authorize @affiliate, :show?
   end
   
   def create_categories
+    authorize Affiliate, :update?
+    
     AffiliateCategory.delete_all affiliate_id: params[:id]
     category_ids = params[:category_ids]
     
@@ -91,14 +89,19 @@ class Admin::System::AffiliatesController < Admin::BaseController
   
   def products
     @affiliate = Affiliate.includes(:products, [products: :product]).find(params[:id])
+    authorize @affiliate, :show?
   end
   
   def create_products
+    authorize Affiliate, :update?
+    
     params[:product_ids].each { |id| AffiliateProduct.create product_id: id, affiliate_id: params[:id] }
     redirect_to :back
   end
   
   def remove_products
+    authorize Affiliate, :update?
+    
     AffiliateProduct.delete_all id: params[:affiliate_product_ids]
     redirect_to :back
   end
